P4: extract _guards.py from helpers.py and migrate consumers
Extract 6 tier/gate guard functions (_require_enabled,
_require_orchestrator_or_higher, _require_orchestrator_exact,
_require_fleet, _check_dry_walkthrough, _validate_skill_command)
from server/helpers.py into server/_guards.py with independent
deferred-import wrappers. Migrate all 13 source-file and 6 test-file
consumers to import from _guards. Remove guard definitions and
orphaned imports (SessionType, session_type, extract_path_arg, os,
gate_error_result, headless_error_result) from helpers.py.
