+++ src/autoskillit/server/tools_ci.py
@@ -12,8 +12,8 @@ from fastmcp.dependencies import CurrentContext
 
 from autoskillit.core import DirectInstall, MarketplaceInstall, get_logger
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
-    _require_enabled,
     _run_subprocess,
     fetch_repo_merge_state,
     resolve_repo_from_remote,
