+++ src/autoskillit/server/tools_ci_watch.py
@@ -14,9 +14,9 @@ from fastmcp.dependencies import CurrentContext
 from autoskillit.core import CIRunScope, get_logger
 from autoskillit.pipeline import ToolContext
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _notify,
-    _require_enabled,
     _run_subprocess,
     resolve_repo_from_remote,
     track_response_size,
