+++ src/autoskillit/server/tools_clone.py
@@ -15,9 +15,9 @@ from fastmcp.dependencies import CurrentContext
 
 from autoskillit.core import CAMPAIGN_ID_ENV_VAR, get_logger
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _notify,
-    _require_enabled,
     clone_registry,
     track_response_size,
 )
