+++ src/autoskillit/server/tools_git.py
@@ -12,9 +12,9 @@ from fastmcp.dependencies import CurrentContext
 
 from autoskillit.core import RestartScope, get_logger
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _notify,
-    _require_enabled,
     _run_subprocess,
     track_response_size,
 )
