+++ src/autoskillit/server/tools_github.py
@@ -15,10 +15,10 @@ from fastmcp.dependencies import CurrentContext
 from autoskillit.core import atomic_write, get_logger
 from autoskillit.pipeline import write_status
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _extract_block,
     _notify,
-    _require_enabled,
     resolve_log_dir,
     track_response_size,
 )
