+++ src/autoskillit/server/tools_status.py
@@ -14,9 +14,9 @@ from fastmcp.dependencies import CurrentContext
 from autoskillit.core import atomic_write, get_logger
 from autoskillit.pipeline import TelemetryFormatter
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _notify,
-    _require_enabled,
     resolve_log_dir,
     track_response_size,
     write_telemetry_clear_marker,
