+++ src/autoskillit/server/tools_workspace.py
@@ -13,9 +13,9 @@ from fastmcp.dependencies import CurrentContext
 
 from autoskillit.core import get_logger, truncate_text
 from autoskillit.server import mcp
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.helpers import (
     _notify,
-    _require_enabled,
     _run_subprocess,
     track_response_size,
 )
