+++ tests/server/test_guards_module.py
@@ -1,0 +1,56 @@
+"""Smoke test: all 6 guards are importable from _guards."""
+
+import pytest
+
+pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
+
+
+def test_require_enabled_importable():
+    from autoskillit.server._guards import _require_enabled
+
+    assert callable(_require_enabled)
+
+
+def test_require_orchestrator_or_higher_importable():
+    from autoskillit.server._guards import _require_orchestrator_or_higher
+
+    assert callable(_require_orchestrator_or_higher)
+
+
+def test_require_orchestrator_exact_importable():
+    from autoskillit.server._guards import _require_orchestrator_exact
+
+    assert callable(_require_orchestrator_exact)
+
+
+def test_require_fleet_importable():
+    from autoskillit.server._guards import _require_fleet
+
+    assert callable(_require_fleet)
+
+
+def test_check_dry_walkthrough_importable():
+    from autoskillit.server._guards import _check_dry_walkthrough
+
+    assert callable(_check_dry_walkthrough)
+
+
+def test_validate_skill_command_importable():
+    from autoskillit.server._guards import _validate_skill_command
+
+    assert callable(_validate_skill_command)
+
+
+def test_guards_not_in_helpers():
+    """After extraction, no guard function should remain in helpers."""
+    import autoskillit.server.helpers as h
+
+    for name in (
+        "_require_enabled",
+        "_require_orchestrator_or_higher",
+        "_require_orchestrator_exact",
+        "_require_fleet",
+        "_check_dry_walkthrough",
+        "_validate_skill_command",
+    ):
+        assert not hasattr(h, name), f"{name} still defined in helpers.py"
