+++ tests/server/test_server_init_gate.py
@@ -10,7 +10,7 @@ import pytest
 import structlog.testing
 
 from autoskillit.pipeline.gate import DefaultGateState
-from autoskillit.server.helpers import _require_enabled
+from autoskillit.server._guards import _require_enabled
 from autoskillit.server.tools_kitchen import _close_kitchen_handler, _open_kitchen_handler
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.medium]
