+++ tests/server/test_server_tool_registration.py
@@ -354,7 +354,7 @@ class TestConfigDrivenBehavior:
     def test_dry_walkthrough_uses_config_marker(self, tool_ctx, tmp_path):
         """S7: Gate checks config.implement_gate.marker."""
         from autoskillit.config import ImplementGateConfig
-        from autoskillit.server.helpers import _check_dry_walkthrough
+        from autoskillit.server._guards import _check_dry_walkthrough
 
         tool_ctx.config = AutomationConfig(
             implement_gate=ImplementGateConfig(marker="CUSTOM MARKER")
@@ -372,7 +372,7 @@ class TestConfigDrivenBehavior:
     def test_dry_walkthrough_uses_config_skill_names(self, tool_ctx, tmp_path):
         """S8: Gate checks config.implement_gate.skill_names."""
         from autoskillit.config import ImplementGateConfig
-        from autoskillit.server.helpers import _check_dry_walkthrough
+        from autoskillit.server._guards import _check_dry_walkthrough
 
         tool_ctx.config = AutomationConfig(
             implement_gate=ImplementGateConfig(skill_names={"/custom-impl"})
