+++ tests/server/test_tools_execution_input_gates.py
@@ -6,10 +6,8 @@ import json
 
 import pytest
 
-from autoskillit.server.helpers import (
-    _check_dry_walkthrough,
-    _get_config,
-)
+from autoskillit.server._guards import _check_dry_walkthrough
+from autoskillit.server.helpers import _get_config
 from autoskillit.server.tools_execution import run_skill
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
@@ -220,17 +218,17 @@ class TestValidateSkillCommand:
     """Unit tests for _validate_skill_command helper."""
 
     def test_returns_none_for_slash_command(self, tool_ctx):
-        from autoskillit.server.helpers import _validate_skill_command
+        from autoskillit.server._guards import _validate_skill_command
 
         assert _validate_skill_command("/autoskillit:investigate") is None
 
     def test_returns_none_for_bare_slash_command(self, tool_ctx):
-        from autoskillit.server.helpers import _validate_skill_command
+        from autoskillit.server._guards import _validate_skill_command
 
         assert _validate_skill_command("/audit-arch") is None
 
     def test_returns_error_json_for_prose(self, tool_ctx):
-        from autoskillit.server.helpers import _validate_skill_command
+        from autoskillit.server._guards import _validate_skill_command
 
         result = _validate_skill_command("Fix the bug")
         assert result is not None
@@ -239,13 +237,13 @@ class TestValidateSkillCommand:
         assert parsed["subtype"] == "gate_error"
 
     def test_returns_error_json_for_empty_string(self, tool_ctx):
-        from autoskillit.server.helpers import _validate_skill_command
+        from autoskillit.server._guards import _validate_skill_command
 
         result = _validate_skill_command("")
         assert result is not None
 
     def test_strips_whitespace_before_check(self, tool_ctx):
-        from autoskillit.server.helpers import _validate_skill_command
+        from autoskillit.server._guards import _validate_skill_command
 
         # Leading whitespace before slash → valid
         assert _validate_skill_command("  /autoskillit:investigate") is None
